Initial commit from CookieCutter template with a few modifications.
Initial configuration followed the cookiecutter installation from [1], but is
modified to drop `.editorconfig`, TravisCI, the default LICENSE, `pbr` (and
therefore `setup.cfg` too, for now), `mypy`, `pytest`, `tox`, and other
testing tools.

In the future, it would be reasonable to bundle those files back in, but for
now it's more important to focus on the essential functionality of this
extension.

[1]: https://github.com/sphinx-contrib/cookiecutter