Adds HTML and LaTeX support for free FontAwesome roles in Sphinx.
Introduces a new role `fa` for font awesome icons.

To use it inline::

	:fa:`check-circle`

To specify one particular style, add it afterwards::

	:fa:`angry solid`

The default style is the first one that appears in the metadata provided by
Font Awesome (usually 'regular').

HTML output is easy enough - for LaTeX output, you will need the `fontawesome`
package installed. Unfortunately, that package is out of date, so not all of
the icons available in FA5 will render correctly. This extension sets the
LaTeX engine to LuaLaTeX. But, some of the icons still render correctly!